Move importStringRange file + test
